Changes the logging level to INFO. (#18)
The previous logging level (`ALL`) displays the following message in the
console when there's a request to `http://localhost:8080/printLogs`:

```
[INFO] GCLOUD: FINEST: Ignoring Exception getting module instance and continuing
[INFO] GCLOUD: com.google.appengine.api.modules.ModulesException: Instance id unavailable
```

The exception is expected and doesn't indicate any problems according to
https://issuetracker.google.com/35896861.

However, the message can be confusing to developers that run the code
sample in this repository. Changing the default logging level to `INFO`
removes the error message from the console.